guidewire-connector/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile 
#
bcrypt==3.1.7             # via paramiko
certifi==2018.8.24        # via requests
cffi==1.11.5              # via bcrypt, cryptography, pynacl
chardet==3.0.4            # via requests
click==6.7                # via flask
cryptography==2.8         # via paramiko, pyopenssl, requests
dnspython==1.15.0         # via eventlet
eventlet==0.24.1
flask-wtf==0.14.2
flask==1.0.2
greenlet==0.4.15          # via eventlet
gunicorn==19.9.0
idna==2.7                 # via requests
itsdangerous==0.24        # via flask
jinja2==2.10              # via flask
markupsafe==1.0           # via jinja2
monotonic==1.5            # via eventlet
paramiko==2.7.1
pycparser==2.18           # via cffi
pyjwt==1.6.4
pynacl==1.3.0             # via paramiko
pyopenssl==18.0.0         # via requests
redis==2.10.6
requests[security]==2.19.1
six==1.11.0               # via bcrypt, cryptography, eventlet, pynacl, pyopenssl
urllib3==1.23             # via requests
werkzeug==0.14.1          # via flask
wtforms==2.2.1            # via flask-wtf
